refactor: move PoF view methods to metaclass
