--- src/ramstk/views/gtk3/pof/workview.pyi
@@ -1,73 +1,0 @@
-# Standard Library Imports
-from typing import Any, Dict, List
-
-# Third Party Imports
-import treelib
-
-# RAMSTK Package Imports
-from ramstk.configuration import RAMSTKUserConfiguration as RAMSTKUserConfiguration
-from ramstk.logger import RAMSTKLogManager as RAMSTKLogManager
-from ramstk.views.gtk3 import GdkPixbuf as GdkPixbuf
-from ramstk.views.gtk3 import Gtk as Gtk
-from ramstk.views.gtk3 import _ as _
-from ramstk.views.gtk3.assistants import AddStressTestMethod as AddStressTestMethod
-from ramstk.views.gtk3.widgets import RAMSTKMessageDialog as RAMSTKMessageDialog
-from ramstk.views.gtk3.widgets import RAMSTKPanel as RAMSTKPanel
-from ramstk.views.gtk3.widgets import RAMSTKWorkView as RAMSTKWorkView
-
-def get_indenture_level(record_id: str) -> str: ...
-
-class PoFPanel(RAMSTKPanel):
-    _dic_column_masks: Dict[str, List[bool]] = ...
-    _dic_headings: Dict[str, List[str]] = ...
-    _module: str = ...
-    _dic_attribute_keys: Any = ...
-    _dic_attribute_updater: Any = ...
-    _dic_row_loader: Any = ...
-    _title: Any = ...
-    dic_damage_models: Any = ...
-    dic_icons: Any = ...
-    dic_load_history: Any = ...
-    dic_measurable_parameters: Any = ...
-    def __init__(self) -> None: ...
-    def do_load_combobox(self) -> None: ...
-    def do_set_callbacks(self) -> None: ...
-    def _on_delete_insert_pof(self, tree: treelib.Tree) -> None: ...
-    _record_id: Any = ...
-    def _on_row_change(self, selection: Gtk.TreeSelection) -> None: ...
-    def __do_load_damage_models(self) -> None: ...
-    def __do_load_load_history(self) -> None: ...
-    def __do_load_measureable_parameters(self) -> None: ...
-    def __do_load_mechanism(
-        self, node: treelib.Node, row: Gtk.TreeIter
-    ) -> Gtk.TreeIter: ...
-    def __do_load_mode(self, node: treelib.Node, row: Gtk.TreeIter) -> Gtk.TreeIter: ...
-    def __do_load_opload(
-        self, node: treelib.Node, row: Gtk.TreeIter
-    ) -> Gtk.TreeIter: ...
-    def __do_load_opstress(
-        self, node: treelib.Node, row: Gtk.TreeIter
-    ) -> Gtk.TreeIter: ...
-    def __do_load_test_method(
-        self, node: treelib.Node, row: Gtk.TreeIter
-    ) -> Gtk.TreeIter: ...
-    def __do_set_properties(self) -> None: ...
-    def __get_cell_model(self, column: Gtk.TreeViewColumn) -> Gtk.TreeModel: ...
-
-class PoF(RAMSTKWorkView):
-    _module: str = ...
-    _pixbuf: bool = ...
-    _tablabel: Any = ...
-    _tabtooltip: Any = ...
-    _lst_tooltips: Any = ...
-    _pnlPanel: Any = ...
-    def __init__(
-        self, configuration: RAMSTKUserConfiguration, logger: RAMSTKLogManager
-    ) -> None: ...
-    def _do_request_delete(self, __button: Gtk.ToolButton) -> None: ...
-    def _do_request_insert_child(self, __button: Gtk.ToolButton) -> None: ...
-    def _do_request_insert_sibling(self, __button: Gtk.ToolButton) -> None: ...
-    _record_id: Any = ...
-    def _do_set_record_id(self, attributes: Dict[str, Any]) -> None: ...
-    def __make_ui(self) -> None: ...
-    def __on_request_insert_opstress_method(self) -> str: ...
